requirements.txt
Flask==3.0.0
Flask-SocketIO==5.3.6
python-socketio==5.10.0
python-engineio==4.8.1
gevent==25.9.1
gevent-websocket==0.10.1
pymongo[srv]==4.6.1
python-dotenv==1.0.0
dnspython==2.4.2
